requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
unsloth>=2024.2.9
peft>=0.9.0
accelerate>=0.24.0
wandb>=0.16.0
python-dotenv>=1.0.0
pyyaml>=6.0